chore: move isartifactory to shared location
