+++ lib/modules/datasource/common.ts
@@ -1,5 +1,6 @@
 import is from '@sindresorhus/is';
 import type { GetPkgReleasesConfig } from './types';
+import type { HttpResponse } from '../../util/http/types';
 
 export function isGetPkgReleasesConfig(
   input: unknown
@@ -12,4 +13,12 @@ export function isGetPkgReleasesConfig(
       (input as GetPkgReleasesConfig).packageName
     )
   );
+}
+
+const JFROG_ARTIFACTORY_RES_HEADER = 'x-jfrog-version';
+
+export function isArtifactoryServer<T = unknown>(
+  res: HttpResponse<T> | undefined
+): boolean {
+  return is.string(res?.headers[JFROG_ARTIFACTORY_RES_HEADER]);
 }
